Automatic license headers verification and application
The current code base lacks license headers and SPDX identifiers in files.

This change uses a Maven plugin to check that all Java files have a license header.

The plugin can be run manually: `mvn license:check` and `mvn license:format`.

The plugin is automatically triggered with the 'release' profile to check for missing headers and apply up-to-date headers.

Note that the plugin maintains copyright year ranges based on Git metadata, which does simplify maintenance.

Signed-off-by: Julien Ponge <jponge@redhat.com>