Optimize Queue usages
This PR improves the sizing and the kind of the various queues we uses.
It creates a central class to create the various type of queues and provide heuristics about the size.

The various operators have been updated to use the right type of queues.

It also fixes #215 as the requests parameter should be used and passed to the operator.