Fix a race condition between onSubscribe / cancel / on{Item,Failure}
There is a possibility for cancel to have the subscription become null because
an item or a failure is emitted.

This removes the dispose() method whose sole purpose was to set the subscription
to null. Since the subscription value being null does not have any value there
is no point keeping it around.