ci: extract the Maven Central deployment as a workflow
Given how unreliable pushing to Maven Central is, we
are better off with a release workflow that makes a source
release, while another (repeatable) workflow can do the
deployment steps. In fact a deployment can fail due to a gateway
error while the artifacts actually end up on Maven Central.