fix: deadlock in Uni memoization
There was a possibility for pipelines having multiple memoized Unis to cause
a deadlock. This is due to the usage of an internal reeentrant lock that was
held even when forwarding upstream and downstream signals, which is bad because
this exposes to deadlocks by design, and this may also involved I/O operations.

Lock access has been rewritten to focus on short, purely state-affecting critical
sections, and signals forwarding happens in a looser fashion.

Note that there is still a cancel() signal that can be forwarded under lock, but
this should be a non-blocking operation as per Reactive Streams semantics and it
shouldn't cause any other signal to back-propagate and deadlock.

Fixes: #1910