fix: race condition in Uni.combine().all()
There was a race condition in UniAndCombination due to concurrent
threads calling AndSupervisor::check, but the item and failure fields
in UniHandler were not volatile, hence the count of incomplete inner
results could be wrong with no further signal to make another check.

An easy fix was to make them volatile, but a work-in-progress guard is
more efficient, and cleaner.

Fixes #1993